+++ linking-platform/link-analytics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/link-analytics
 
+## 1.0.0
+
+### Major Changes
+
+- [`e7c31586ebe`](https://bitbucket.org/atlassian/atlassian-frontend/commits/e7c31586ebe) - Add support for deriving create link attributes from another event.
+
 ## 0.1.1
 
 ### Patch Changes